Use PlonePAS methods for managing groups.
These methods are PAS-aware and thus work with other PAS plugins that handle groups. Thanks to Giovanni Toffoli for the initial patch. This fix differs from the initial patch in that it also includes CSRF protections, and the getGroupMembers result is converted to member objects (new method only returns user ids).

This fixes #9163.

svn path=/Plone/branches/3.3/; revision=30837